테스트 모델링 추가 및 총 발사 메커니즘 변경

+++ Assets/Scripts/Taeseung/Weapon/Gun/Interface/WeaponInterface.cs
@@ -19,11 +19,11 @@ public interface WeaponInterface
     /// 게이지 셋팅
     /// </summary>
     /// <param name="newval"></param>
-    public void SetWeaponGauge(float newval);
+    public void SetWeaponGauge(int newval);
 
     /// <summary>
     /// 게이지량 가져오기
     /// </summary>
     /// <returns></returns>
-    public float GetWeaponGauge();
+    public int GetWeaponGauge();
 }
